requirements.txt
Flask==3.0.3
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==3.0.3
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
idna==3.7
itsdangerous==2.2.0
lxml==5.3.0
mailchimp-marketing==3.0.80
numpy==2.0.1
panda==0.3.1
pandas==2.2.2
pillow==10.4.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Flask-WTF==1.2.1
sqlalchemy==2.0.32
pymysql==1.1.1
flask-jwt-extended==4.6.0